requirements.txt
discord.py==1.5.1
tortoise-orm==0.16.18